Update train_verdict_predictor.py; sentence_transformers not used and not in requirements.
Hello, I was using your library, and I had a problem with from sentence_transformers import CrossEncoder since sentence_transformers is not specified in the requirements. I think it's safe to remove it because it's not used, just imported, and to avoid installation problems, it's better to remove it as it's not declared in the requirements.